Criado: Configuração do mockito, criação de mocks, injeção de mocks
